+++ Classes/AdminPanel/SchemaModule.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace Brotkrueml\Schema\AdminPanel;
 
-use Brotkrueml\Schema\Cache\PagesCacheService;
+use Brotkrueml\Schema\Cache\MarkupCacheHandler;
 use Brotkrueml\Schema\Extension;
 use Symfony\Component\DependencyInjection\Attribute\Autoconfigure;
 use TYPO3\CMS\Adminpanel\ModuleApi\AbstractModule;
@@ -24,7 +24,7 @@ use TYPO3\CMS\Adminpanel\ModuleApi\ShortInfoProviderInterface;
 final class SchemaModule extends AbstractModule implements ShortInfoProviderInterface
 {
     public function __construct(
-        private readonly PagesCacheService $pagesCacheService,
+        private readonly MarkupCacheHandler $markupCacheHandler,
     ) {}
 
     public function getIconIdentifier(): string
@@ -44,7 +44,7 @@ final class SchemaModule extends AbstractModule implements ShortInfoProviderInte
 
     public function getShortInfo(): string
     {
-        $jsonLd = $this->pagesCacheService->getMarkupFromCache() ?? '';
+        $jsonLd = $this->markupCacheHandler->getMarkup() ?? '';
 
         $numberOfTypes = 0;
         if ($jsonLd !== '') {