+++ Tests/Unit/AdminPanel/TypesInformationTest.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 namespace Brotkrueml\Schema\Tests\Unit\AdminPanel;
 
 use Brotkrueml\Schema\AdminPanel\TypesInformation;
-use Brotkrueml\Schema\Cache\PagesCacheService;
+use Brotkrueml\Schema\Cache\MarkupCacheHandler;
 use Brotkrueml\Schema\Extension;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -28,16 +28,16 @@ use TYPO3\CMS\Fluid\View\StandaloneView;
 #[CoversClass(TypesInformation::class)]
 final class TypesInformationTest extends TestCase
 {
-    private PagesCacheService&Stub $pagesCacheServiceStub;
+    private MarkupCacheHandler&Stub $markupCacheHandlerStub;
     private MockObject $viewMock;
     private LanguageService&Stub $languageServiceStub;
     private TypesInformation $subject;
 
     protected function setUp(): void
     {
-        $this->pagesCacheServiceStub = self::createStub(PagesCacheService::class);
+        $this->markupCacheHandlerStub = self::createStub(MarkupCacheHandler::class);
 
-        $this->subject = new TypesInformation($this->pagesCacheServiceStub);
+        $this->subject = new TypesInformation($this->markupCacheHandlerStub);
 
         $this->viewMock = $this->createMock(StandaloneView::class);
         $this->viewMock
@@ -76,8 +76,8 @@ final class TypesInformationTest extends TestCase
     #[DataProvider('dataProviderForGetContent')]
     public function getContentWithNoCacheEntryAvailable(?string $markupFromCache, array $expectedTypes): void
     {
-        $this->pagesCacheServiceStub
-            ->method('getMarkupFromCache')
+        $this->markupCacheHandlerStub
+            ->method('getMarkup')
             ->willReturn($markupFromCache ? \sprintf(Extension::JSONLD_TEMPLATE, $markupFromCache) : $markupFromCache);
 
         $this->viewMock